@@ -12,14 +12,15 @@ struct GetUsers: Endpoint {
     var page: Int = 1
     var recordsPerPage: Int = 10
     var name: String? = nil
-    
-    var call: some HTTPCall {
-        GoRESTCall()
-            .path("/public/v2/users")
+
+    var request: some Request<[User]> {
+        Get<[User]>("users")
             .queries {
                 Query(name: "page", value: String(page))
                 Query(name: "per_page", value: String(recordsPerPage))
-                Query(name: "name", value: name)
+                if let name {
+                    Query(name: "name", value: name)
+                }
             }
     }
 }
